API_Mosquitto/API/requirements.txt
fastapi
paho-mqtt
uvicorn